moved common fixtures to tests/runtime/conftest.py
